feat(templating): remove cycle
remove cycle between view-engine, html-behavior, and module-analyzer by
combining into view-engine.

also update references in decorators and index
